move mock VDR to mocks pkg
