+++ validation/verifiable_presentation.go
@@ -3,13 +3,19 @@ package validation
 import (
 	"fmt"
 
+	"github.com/hyperledger/aries-framework-go/pkg/doc/did"
+
 	"github.com/hyperledger/aries-framework-go/pkg/framework/aries/api/vdr"
 
 	"github.com/medibloc/vc-sdk/pkg/vc"
 )
 
+type didResolver interface {
+	Resolve(did string, opts ...vdr.DIDMethodOption) (*did.DocResolution, error)
+}
+
 // ValidateVP validates verifiable presentation
-func ValidateVP(vdr vdr.Registry, vpBytes, pdBytes []byte) error {
+func ValidateVP(vdr didResolver, vpBytes, pdBytes []byte) error {
 	f, err := vc.NewFramework(vdr)
 	if err != nil {
 		return fmt.Errorf("failed to create a framework for VP verification: %w", err)
